Texture and Compute Shader abstraction
